backend/requirements.txt
# Web Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Web Scraping
selenium>=4.16.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# RAG & LLM (compatible versions for Python 3.9)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.1.0
openai>=1.12.0
chromadb>=0.4.0,<0.5.0
tiktoken>=0.5.0
numpy<2.0.0,>=1.22.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-json-logger>=2.0.0

# UI
streamlit>=1.30.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0

# Testing
colorama>=0.4.6

# Caching & Performance
redis>=5.0.0,<6.0.0
hiredis>=2.2.0  # Fast Redis client
cachetools>=5.3.0  # In-memory LRU cache

# Load Testing
aiohttp>=3.9.0  # Async HTTP client
